+++ tests/Public/Start-GuestConfigurationRemediation.Tests.ps1
@@ -97,4 +97,45 @@ Describe 'Start-GuestConfigurationPackageRemediation' {
             $fileContent | Should -Be $expectedContent
         }
     }
+
+    Context 'TestFile package with parameters - no metaconfig' {
+        BeforeAll {
+            $script:testFilePackagePath = Join-Path -Path $testPackagesFolderPath -ChildPath 'TestFilePackage_nometa.zip'
+            $testFilePath = Join-Path -Path $TestDrive -ChildPath 'Mordor.txt'
+            $expectedContent = 'Frodo Baggins'
+
+            if (Test-Path -Path $testFilePath)
+            {
+                $null = Remove-Item -Path $testFilePath -Force
+            }
+
+            $parameters = @(
+                @{
+                    ResourceType = 'TestFile'
+                    ResourceId = 'MyTestFile'
+                    ResourcePropertyName = 'Path'
+                    ResourcePropertyValue = $testFilePath
+                },
+                @{
+                    ResourceType = 'TestFile'
+                    ResourceId = 'MyTestFile'
+                    ResourcePropertyName = 'Content'
+                    ResourcePropertyValue = $expectedContent
+                }
+            )
+        }
+
+        It 'Set should run without throwing' {
+            { Start-GuestConfigurationPackageRemediation -Path $script:testFilePackagePath -Parameter $parameters } | Should -Not -Throw
+        }
+
+        It 'Test file should exist at expected path' {
+            Test-Path -Path $testFilePath | Should -Be $True
+        }
+
+        It 'Test file should have expected content' {
+            $fileContent = Get-Content -Path $testFilePath -Raw
+            $fileContent | Should -Be $expectedContent
+        }
+    }
 }