+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0-preview0008]
+
+## Fixed 
+
+- Start-GuestConfigurationPackageComplianceStatus and Start-GuestConfigurationPackageRemediation will run again on packages that are missing the metaconfig file. Added tests for this as well.
+- Fixed some typos in the New-GuestConfigurationPolicy help comments
+
 ## [4.0.0-preview0007] - 2022-05-19
 
 ### Changed